+++ packages/library-sync/src/jvm/kotlin/io/realm/mongodb/internal/HttpClientCache.kt
@@ -11,9 +11,12 @@ import io.ktor.client.features.logging.Logger
  * https://github.com/realm/realm-kotlin/issues/480 only seem to be a problem on macOS.
  */
 internal actual class HttpClientCache actual constructor(timeoutMs: Long, customLogger: Logger?) {
-    private val client = createClient(timeoutMs, customLogger)
+    private val httpClient: HttpClient by lazy { createClient(timeoutMs, customLogger) }
     actual fun getClient(): HttpClient {
-        return client
+        return httpClient
+    }
+    actual fun close() {
+        httpClient.close()
     }
 }