+++ test/base/src/androidTest/kotlin/io/realm/test/shared/RealmListTests.kt
@@ -35,6 +35,9 @@ import io.realm.realmListOf
 import io.realm.test.platform.PlatformUtils
 import io.realm.test.util.TypeDescriptor
 import io.realm.toRealmList
+import kotlinx.coroutines.async
+import kotlinx.coroutines.flow.takeWhile
+import kotlinx.coroutines.runBlocking
 import kotlin.reflect.KClassifier
 import kotlin.reflect.KMutableProperty1
 import kotlin.test.AfterTest
@@ -342,6 +345,25 @@ class RealmListTests {
             }
         }
         assertEquals(4, realm.query<Sample>().find().size)
+    }
+
+    @Test
+    fun listNotifications() = runBlocking {
+        val container = realm.writeBlocking { copyToRealm(RealmListContainer()) }
+        val collect = async {
+            container.objectListField.asFlow()
+                .takeWhile { it.list.size < 5 }
+                .collect {
+                    it.list.forEach {
+                        // No-op ... just verifying that we can access each element. See https://github.com/realm/realm-kotlin/issues/827
+                    }
+                }
+        }
+        while (!collect.isCompleted) {
+            realm.writeBlocking {
+                findLatest(container)!!.objectListField.add(RealmListContainer())
+            }
+        }
     }
 
     private fun getCloseableRealm(): Realm =