+++ test/base/src/commonMain/kotlin/io/realm/entities/embedded/EmbeddedInnerChild.kt
@@ -2,6 +2,6 @@ package io.realm.entities.embedded
 
 import io.realm.EmbeddedObject
 
-class EmbeddedInnerChild: EmbeddedObject{
+class EmbeddedInnerChild : EmbeddedObject {
     var id: String? = "Embedded child"
 }